enable animations during rendering
I should have changed it step by step to get proper commits, but I had to rush the implementation somewhat as I do not have a lot of time at the moment :(
The concept of a camera has changed, it now acts as a quadratic texture. Thus, the jittering etc. had to be moved to the renderer. However, the camera does no longer need to know anything about resolution or aspect ratio.
The images are now also saved from the renderer as an animation will consist of several images that all need to be saved. Stepping through the animation is done in the renderer by calling the step() function on the scene.
The scene is now created by a scene_builder because it consists of several keyframes between which the geometry and the camera will be interpolated.